pyproject.toml
[project]
name = "banrepco-api"
version = "0.1.0"
description = "Useful catalog of APIs for the Bank of the Republic of Colombia"
authors = [{ name = "Kevin Rodríguez", email = "kyrodriguez97@gmail.com" }]
requires-python = ">=3.13,<3.15"
readme = "README.md"
dependencies = [
    "alembic==1.18.4",
    "fastapi-pagination==0.15.12",
    "fastapi[standard]==0.136.1",
    "sqlalchemy-libsql==0.2.0",
    "uvicorn==0.46.0",
    "pydantic==2.13.3",
    "schedule==1.2.2",
    "requests==2.33.1",
    "xmltodict==1.0.4",
]

[dependency-groups]
test = [
    "pytest==9.0.3",
    "pytest-env==1.6.0",
    "pytest-cov==7.1.0",
]
dev = [
    "pre-commit==4.6.0",
    "isort==7.0.0",
    "black==25.9.0",
    "flake8==7.3.0",
]

[tool.uv]
package = false
